requirements.txt
openai==0.28.0
streamlit 
python-dotenv
